- restored a fraction of the similarsearch glory. Still unreliable - HashSmallestDiffPCT will be returned as float in case precision is needed

+++ Trunk/GameScripts/pythontests/imagesearchdll.py
@@ -82,7 +82,7 @@ def _parse_single_result(raw: str) -> SingleResult:
             avg_color_diff  = int(toks[5]) if len(toks) > 5 else -1,
             color_diff_count= int(toks[6]) if len(toks) > 6 else -1,
             color_diff_pct  = int(toks[7]) if len(toks) > 7 else -1,
-            hash_diff_pct   = int(toks[8]) if len(toks) > 8 else -1,
+            hash_diff_pct   = float(toks[8]) if len(toks) > 8 else -1,
             satd            = int(toks[9]) if len(toks) > 9 else -1,
             satd_per_pixel   = int(toks[10]) if len(toks) > 10 else -1,
             sad_brightness_corrected   = int(toks[11]) if len(toks) > 11 else -1,
